fix: reparent ROUTES type to routing/types
