+++ src/components/CheckEditor/CheckProbes/PrivateProbesAlert.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Alert, LinkButton, Stack, TextLink } from '@grafana/ui';
 import { useLocalStorage } from 'usehooks-ts';
 
-import { ROUTES } from 'types';
+import { ROUTES } from 'routing/types';
 import { getRoute } from 'routing/utils';
 
 export const PrivateProbesAlert = () => {
