+++ src/components/CheckList/CheckList.test.tsx
@@ -14,7 +14,8 @@ import { render } from 'test/render';
 import { server } from 'test/server';
 import { getSelect, selectOption } from 'test/utils';
 
-import { Check, FeatureName, ROUTES } from 'types';
+import { Check, FeatureName } from 'types';
+import { ROUTES } from 'routing/types';
 import { generateRoutePath } from 'routing/utils';
 
 import { CheckList } from './CheckList';
