+++ src/components/CheckListItem/CheckItemActionButtons.tsx
@@ -3,7 +3,8 @@ import { GrafanaTheme2 } from '@grafana/data';
 import { ConfirmModal, IconButton, LinkButton, useStyles2 } from '@grafana/ui';
 import { css } from '@emotion/css';
 
-import { Check, ROUTES } from 'types';
+import { Check } from 'types';
+import { ROUTES } from 'routing/types';
 import { generateRoutePath, getRoute } from 'routing/utils';
 import { useDeleteCheck } from 'data/useChecks';
 import { useCanReadMetrics, useCanWriteSM } from 'hooks/useDSPermission';
