+++ src/components/OverLimitAlert/ExecutionLimitAlert.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { TextLink } from '@grafana/ui';
 
-import { ROUTES } from 'types';
+import { ROUTES } from 'routing/types';
 import { getRoute } from 'routing/utils';
 import { FREE_EXECUTION_LIMIT } from 'hooks/useAtHgExecutionLimit';
 import { Ul } from 'components/Ul';
