+++ src/components/ProbeEditor/ProbeEditor.tsx
@@ -6,7 +6,8 @@ import { css } from '@emotion/css';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { ProbeSchema } from 'schemas/forms/ProbeSchema';
 
-import { ExtendedProbe, FeatureName, Probe, ROUTES } from 'types';
+import { ExtendedProbe, FeatureName, Probe } from 'types';
+import { ROUTES } from 'routing/types';
 import { getRoute } from 'routing/utils';
 import { useCanEditProbe } from 'hooks/useCanEditProbe';
 import { FeatureFlag } from 'components/FeatureFlag';
