+++ src/components/SceneRedirecter.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Navigate } from 'react-router-dom-v5-compat';
 import { LoadingPlaceholder } from '@grafana/ui';
 
-import { ROUTES } from 'types';
+import { ROUTES } from 'routing/types';
 import { generateRoutePath } from 'routing/utils';
 import { useChecks } from 'data/useChecks';
 import { useQuery } from 'hooks/useQuery';
