+++ src/hooks/useCheckTypeGroupOptions.tsx
@@ -1,7 +1,8 @@
 import { ReactNode, useContext } from 'react';
 import { IconName } from '@grafana/data';
 
-import { CheckType, CheckTypeGroup, FeatureName, ROUTES } from 'types';
+import { CheckType, CheckTypeGroup, FeatureName } from 'types';
+import { ROUTES } from 'routing/types';
 import { getRoute } from 'routing/utils';
 import { FeatureFlagContext } from 'contexts/FeatureFlagContext';
