+++ src/page/ChecksPage.test.tsx
@@ -10,9 +10,10 @@ import { apiRoute } from 'test/handlers';
 import { ComponentWrapperProps, render } from 'test/render';
 import { server } from 'test/server';
 
-import { AlertSensitivity, Check, CheckTypeGroup, ROUTES } from 'types';
+import { AlertSensitivity, Check, CheckTypeGroup } from 'types';
 import { PLUGIN_URL_PATH } from 'routing/constants';
 import { InitialisedRouter } from 'routing/InitialisedRouter';
+import { ROUTES } from 'routing/types';
 import { generateRoutePath } from 'routing/utils';
 
 import { FeatureFlagProvider } from '../components/FeatureFlagProvider';
