+++ src/page/ConfigPageLayout/ConfigPageLayout.test.tsx
@@ -3,7 +3,7 @@ import { MemoryRouter } from 'react-router-dom';
 import { CompatRouter, Route, Routes } from 'react-router-dom-v5-compat';
 import { render } from '@testing-library/react';
 
-import { ROUTES } from '../../types';
+import { ROUTES } from 'routing/types';
 import { getRoute } from 'routing/utils';
 
 import { DataTestIds } from '../../test/dataTestIds';
