+++ src/page/DashboardPage.tsx
@@ -3,8 +3,9 @@ import { useParams } from 'react-router-dom-v5-compat';
 import { SceneApp, SceneAppPage } from '@grafana/scenes';
 import { Spinner } from '@grafana/ui';
 
-import { CheckPageParams, CheckType, DashboardSceneAppConfig, FeatureName, ROUTES } from 'types';
+import { CheckPageParams, CheckType, DashboardSceneAppConfig, FeatureName } from 'types';
 import { getCheckType } from 'utils';
+import { ROUTES } from 'routing/types';
 import { generateRoutePath } from 'routing/utils';
 import { useChecks } from 'data/useChecks';
 import { useFeatureFlag } from 'hooks/useFeatureFlag';
