+++ src/page/EditProbe/EditProbe.test.tsx
@@ -5,8 +5,9 @@ import { apiRoute, getServerRequests } from 'test/handlers';
 import { render } from 'test/render';
 import { server } from 'test/server';
 
-import { Probe, ROUTES } from 'types';
+import { Probe } from 'types';
 import { formatDate } from 'utils';
+import { ROUTES } from 'routing/types';
 import { generateRoutePath, getRoute } from 'routing/utils';
 
 import { DataTestIds } from '../../test/dataTestIds';
