+++ src/page/NewCheck/NewCheck.tsx
@@ -2,7 +2,8 @@ import React from 'react';
 import { useParams } from 'react-router-dom-v5-compat';
 import { TextLink } from '@grafana/ui';
 
-import { CheckFormPageParams, ROUTES } from 'types';
+import { CheckFormPageParams } from 'types';
+import { ROUTES } from 'routing/types';
 import { getRoute } from 'routing/utils';
 import { CHECK_TYPE_GROUP_OPTIONS } from 'hooks/useCheckTypeGroupOptions';
 import { CheckForm } from 'components/CheckForm/CheckForm';
