+++ src/page/NewProbe/NewProbe.test.tsx
@@ -5,7 +5,8 @@ import { ADD_PROBE_TOKEN_RESPONSE } from 'test/fixtures/probes';
 import { render } from 'test/render';
 import { fillProbeForm } from 'test/utils';
 
-import { FeatureName, ROUTES } from 'types';
+import { FeatureName } from 'types';
+import { ROUTES } from 'routing/types';
 import { generateRoutePath, getRoute } from 'routing/utils';
 
 import { DataTestIds } from '../../test/dataTestIds';
