+++ src/page/SubsectionWelcomePage.tsx
@@ -4,7 +4,7 @@ import { PluginPage } from '@grafana/runtime';
 import { Stack, Text, useStyles2 } from '@grafana/ui';
 import { css } from '@emotion/css';
 
-import { ROUTES } from 'types';
+import { ROUTES } from 'routing/types';
 import { AppInitializer } from 'components/AppInitializer';
 import { Card } from 'components/Card';
