+++ src/page/__testHelpers__/checkForm.tsx
@@ -6,8 +6,9 @@ import { apiRoute, getServerRequests } from 'test/handlers';
 import { render } from 'test/render';
 import { server } from 'test/server';
 
-import { Check, CheckType, ROUTES } from 'types';
+import { Check, CheckType } from 'types';
 import { getCheckTypeGroup } from 'utils';
+import { ROUTES } from 'routing/types';
 import { generateRoutePath, getRoute } from 'routing/utils';
 import { EditCheck } from 'page/EditCheck';
 import { NewCheck } from 'page/NewCheck';
