+++ src/routing/InitialisedRouter.test.tsx
@@ -3,9 +3,9 @@ import { screen } from '@testing-library/react';
 import { SM_DATASOURCE } from 'test/fixtures/datasources';
 import { type CustomRenderOptions, render } from 'test/render';
 
-import { ROUTES } from 'types';
 import { PLUGIN_URL_PATH } from 'routing/constants';
 import { InitialisedRouter } from 'routing/InitialisedRouter';
+import { ROUTES } from 'routing/types';
 import { getRoute } from 'routing/utils';
 
 function renderInitialisedRouting(options?: CustomRenderOptions) {
