+++ src/routing/LegacyEditRedirect.tsx
@@ -1,6 +1,6 @@
 import { useNavigate, useParams } from 'react-router-dom-v5-compat';
 
-import { ROUTES } from 'types';
+import { ROUTES } from 'routing/types';
 import { generateRoutePath } from 'routing/utils';
 
 interface LegacyEditRedirectProps {
