+++ src/routing/types.ts
@@ -1,0 +1,16 @@
+export enum ROUTES {
+  Alerts = 'alerts',
+  CheckDashboard = 'checks/:id',
+  Checks = 'checks',
+  ChooseCheckGroup = 'checks/choose-type',
+  Config = 'config', // config (index)
+  EditCheck = 'checks/:id/edit',
+  ViewProbe = 'probes/:id',
+  EditProbe = 'probes/:id/edit',
+  Home = 'home',
+  NewCheck = 'checks/new',
+  NewProbe = 'probes/new',
+  Probes = 'probes',
+  Redirect = 'redirect',
+  Scene = 'scene',
+}
