+++ src/types.ts
@@ -675,23 +675,6 @@ export interface UsageValues {
   dpm: number;
 }
 
-export enum ROUTES {
-  Alerts = 'alerts',
-  CheckDashboard = 'checks/:id',
-  Checks = 'checks',
-  ChooseCheckGroup = 'checks/choose-type',
-  Config = 'config', // config (index)
-  EditCheck = 'checks/:id/edit',
-  ViewProbe = 'probes/:id',
-  EditProbe = 'probes/:id/edit',
-  Home = 'home',
-  NewCheck = 'checks/new',
-  NewProbe = 'probes/new',
-  Probes = 'probes',
-  Redirect = 'redirect',
-  Scene = 'scene',
-}
-
 interface Params extends Record<string, string | undefined> {}
 
 export interface CheckPageParams extends Params {
